requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
# Editable install with no version control (thedirector==0.1.0)
-e .
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==7.34.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
slack_sdk==3.41.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
